+++ routers/user.routing.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router()
-const userController = require('../controllers/userController')
+const userController = require('../controllers/user.controller')
 
 router.post('', userController.createUser)
 router.put('/:id', userController.updateUser)
